# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Blog API",
    "description": "A microservices-based blog application API built with Go, implementing CQRS pattern with RabbitMQ for asynchronous message processing.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    }
  ],
  "tags": [
    {
      "name": "Health",
      "description": "Health check endpoints"
    },
    {
      "name": "Posts",
      "description": "Post management endpoints"
    },
    {
      "name": "Authentication",
      "description": "OAuth authentication endpoints"
    }
  ],
  "paths": {
    "/ping": {
      "get": {
        "tags": ["Health"],
        "summary": "Health check",
        "description": "Returns a simple health check response",
        "operationId": "ping",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PingResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/posts": {
      "get": {
        "tags": ["Posts"],
        "summary": "List posts",
        "description": "Retrieve paginated posts, optionally filtered by slug, text content, or author",
        "operationId": "listPosts",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number (1-based)",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "Number of items per page",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "slug",
            "in": "query",
            "description": "Filter posts by slug (partial match)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "text",
            "in": "query",
            "description": "Search posts by content text (partial match)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "author",
            "in": "query",
            "description": "Filter posts by author name (partial match)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedPostResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "tags": ["Posts"],
        "summary": "Create post",
        "description": "Create a new blog post. The operation is processed asynchronously via RabbitMQ.",
        "operationId": "createPost",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Post creation accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/posts/{id}": {
      "get": {
        "tags": ["Posts"],
        "summary": "Get post by ID",
        "description": "Retrieve a single post by its UUID",
        "operationId": "getPostById",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Post UUID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "tags": ["Posts"],
        "summary": "Delete post",
        "description": "Delete a post by its UUID. The operation is processed asynchronously via RabbitMQ.",
        "operationId": "deletePost",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Post UUID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Post deletion accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/auth/{provider}": {
      "get": {
        "tags": ["Authentication"],
        "summary": "OAuth login",
        "description": "Initiate OAuth authentication flow with the specified provider. Redirects to OAuth provider for authentication. If the user is already authenticated, they will be redirected to CLIENT_URL.",
        "operationId": "oauthLogin",
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "description": "OAuth provider (e.g., github)",
            "schema": {
              "type": "string",
              "enum": ["github"]
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to OAuth provider or CLIENT_URL if already authenticated",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                },
                "description": "Redirect URL"
              }
            }
          }
        }
      }
    },
    "/auth/{provider}/callback": {
      "get": {
        "tags": ["Authentication"],
        "summary": "OAuth callback",
        "description": "OAuth callback endpoint that handles the authentication response from the provider. Automatically creates a user account if one doesn't exist. User creation is processed asynchronously via RabbitMQ. A session is created and stored in a cookie.",
        "operationId": "oauthCallback",
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "description": "OAuth provider (e.g., github)",
            "schema": {
              "type": "string",
              "enum": ["github"]
            }
          },
          {
            "name": "code",
            "in": "query",
            "required": false,
            "description": "OAuth authorization code",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "description": "OAuth state parameter",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to CLIENT_URL after successful authentication",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                },
                "description": "Redirect URL (CLIENT_URL)"
              }
            }
          },
          "500": {
            "description": "Error authenticating user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout/{provider}": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Logout",
        "description": "Log out the user, clear the session, and redirect to home",
        "operationId": "logout",
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "description": "OAuth provider (e.g., github)",
            "schema": {
              "type": "string",
              "enum": ["github"]
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to home",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                },
                "description": "Redirect URL (/)"
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "blog_session",
        "description": "Session cookie set after successful OAuth authentication"
      }
    },
    "schemas": {
      "Post": {
        "type": "object",
        "required": ["id", "slug", "title", "content", "author"],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Post UUID"
          },
          "slug": {
            "type": "string",
            "minLength": 3,
            "maxLength": 255,
            "description": "SEO-friendly URL slug"
          },
          "title": {
            "type": "string",
            "minLength": 3,
            "maxLength": 255,
            "description": "Post title"
          },
          "content": {
            "type": "string",
            "minLength": 10,
            "maxLength": 10000,
            "description": "Post content"
          },
          "author": {
            "type": "string",
            "minLength": 3,
            "maxLength": 255,
            "description": "Author name"
          }
        }
      },
      "PaginatedPostResponse": {
        "type": "object",
        "required": ["items", "total", "page", "page_size"],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Post"
            },
            "description": "Array of posts"
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "description": "Total number of posts matching the filters"
          },
          "page": {
            "type": "integer",
            "description": "Current page number"
          },
          "page_size": {
            "type": "integer",
            "description": "Number of items per page"
          }
        }
      },
      "CreatePostRequest": {
        "type": "object",
        "required": ["Id", "Slug", "Title", "Content", "Author"],
        "properties": {
          "Id": {
            "type": "string",
            "format": "uuid",
            "description": "Post UUID"
          },
          "Slug": {
            "type": "string",
            "minLength": 3,
            "maxLength": 255,
            "pattern": "^[a-zA-Z0-9]+$",
            "description": "SEO-friendly URL slug (alphanumeric only)"
          },
          "Title": {
            "type": "string",
            "minLength": 3,
            "maxLength": 255,
            "description": "Post title"
          },
          "Content": {
            "type": "string",
            "minLength": 10,
            "maxLength": 10000,
            "description": "Post content"
          },
          "Author": {
            "type": "string",
            "minLength": 3,
            "maxLength": 255,
            "description": "Author name"
          }
        }
      },
      "PingResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "pong"
          }
        }
      },
      "MessageResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "error": {
            "type": "string"
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized - Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "message": {
                  "type": "string",
                  "example": "Unauthorized User"
                },
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  }
}